requirements.txt
Flask==0.12.2
textblob==0.12.0
python-twitter==3.3
pycorenlp
pattern